adicionando endpoints livro/info

+++ services/autor.service.js
@@ -1,4 +1,5 @@
 import AutorRepository from "../repositories/autor.repository.js";
+import LivroRepository from "../repositories/livro.repository.js";
 
 async function createAutor(autor) {
     return await AutorRepository.insertAutor(autor);
@@ -9,6 +10,10 @@ async function updateAutor(autor) {
 };
 
 async function deleteAutor(id) {
+    const livros = await LivroRepository.getLivroByAutorId(id);
+    if(livros.length > 0) {
+        throw new Error("Não é possível deletar esse autor pois o mesmo tem um livro");
+    }
     await AutorRepository.deleteAutor(id);
 };
